Removed unused code from CrcHelper class.
The byte order of the result of the calculation now matches the byte order used by the System.BitConverter class.
Added CrcTest class to verify the correctness of algorithm calculations.
The Crc class IsRight method now does not throw an exception if the calculated value is different from the test value.